fix: incorrect type of crc32; progress bar increasement

+++ struct.go
@@ -33,7 +33,7 @@ type prepareSendResp struct {
 
 type uploadResponse struct {
 	Ticket string `json:"ctx"`
-	Hash   int    `json:"crc32"`
+	Hash   int64  `json:"crc32"`
 }
 
 type beforeSendResp struct {
